GH-5 : Moving link error enum into enums package
